Prevent selection of an empty range when editing activity
