+++ web/admin/components/C1BExport.js
@@ -21,6 +21,7 @@ import Switch from "@material-ui/core/Switch/Switch";
 import { DAY, isoFormatLocalDate } from "common/utils/time";
 import Alert from "@material-ui/lab/Alert";
 import { HTTP_QUERIES } from "common/utils/apiQueries";
+import { DateOrDateTimeRangeSelectionContext } from "common/components/DateOrDateTimeRangeSelectionContext";
 
 const useStyles = makeStyles(theme => ({
   start: {
@@ -73,14 +74,6 @@ export default function C1BExport({
       );
     } else setDateRangeError(null);
   }, [minDate, maxDate]);
-
-  React.useEffect(() => {
-    if (minDate && maxDate && maxDate < minDate) setMaxDate(minDate);
-  }, [minDate]);
-
-  React.useEffect(() => {
-    if (maxDate && minDate && minDate > maxDate) setMinDate(maxDate);
-  }, [maxDate]);
 
   const [_companies, setCompanies] = React.useState([]);
   const [_users, setUsers] = React.useState([]);
@@ -165,42 +158,46 @@ export default function C1BExport({
           >
             <EmployeeFilter users={_users} setUsers={setUsers} />
           </Grid>
-          <Grid item sm={6}>
-            <DatePicker
-              required
-              label="Date de début"
-              value={minDate}
-              format="d MMMM yyyy"
-              onChange={setMinDate}
-              clearable
-              cancelLabel={null}
-              clearLabel="Annuler"
-              autoOk
-              disableFuture
-              inputVariant="outlined"
-              animateYearScrolling
-              error={!!dateRangeError}
-              helperText={dateRangeError}
-            />
-          </Grid>
-          <Grid item sm={6}>
-            <DatePicker
-              required
-              label="Date de fin"
-              value={maxDate}
-              format="d MMMM yyyy"
-              onChange={setMaxDate}
-              clearable
-              cancelLabel={null}
-              clearLabel="Annuler"
-              autoOk
-              disableFuture
-              inputVariant="outlined"
-              animateYearScrolling
-              error={!!dateRangeError}
-              helperText={dateRangeError}
-            />
-          </Grid>
+          <DateOrDateTimeRangeSelectionContext
+            start={minDate}
+            setStart={setMinDate}
+            end={maxDate}
+            setEnd={setMaxDate}
+            nullableBounds={false}
+          >
+            <Grid item sm={6}>
+              <DatePicker
+                required
+                label="Date de début"
+                value={minDate}
+                format="d MMMM yyyy"
+                onChange={setMinDate}
+                cancelLabel={null}
+                autoOk
+                disableFuture
+                inputVariant="outlined"
+                animateYearScrolling
+                error={!!dateRangeError}
+                helperText={dateRangeError}
+              />
+            </Grid>
+            <Grid item sm={6}>
+              <DatePicker
+                required
+                label="Date de fin"
+                value={maxDate}
+                format="d MMMM yyyy"
+                onChange={setMaxDate}
+                cancelLabel={null}
+                autoOk
+                disableFuture
+                inputVariant="outlined"
+                animateYearScrolling
+                error={!!dateRangeError}
+                helperText={dateRangeError}
+              />
+            </Grid>
+          </DateOrDateTimeRangeSelectionContext>
           <Grid item xs={12}>
             <Box className={classes.switchContainer}>
               <Switch
