+++ web/pwa/components/PDFExport.js
@@ -22,6 +22,7 @@ import {
   startOfMonthAsDate
 } from "common/utils/time";
 import { HTTP_QUERIES } from "common/utils/apiQueries";
+import { DateOrDateTimeRangeSelectionContext } from "common/components/DateOrDateTimeRangeSelectionContext";
 
 const useStyles = makeStyles(theme => ({
   start: {
@@ -67,14 +68,6 @@ export default function PDFExport({ open, handleClose }) {
     } else setDateRangeError(null);
   }, [minDate, maxDate]);
 
-  React.useEffect(() => {
-    if (minDate && (!maxDate || maxDate < minDate)) setMaxDate(minDate);
-  }, [minDate]);
-
-  React.useEffect(() => {
-    if (maxDate && (!minDate || minDate > maxDate)) setMinDate(maxDate);
-  }, [maxDate]);
-
   const classes = useStyles();
 
   return (
@@ -90,53 +83,55 @@ export default function PDFExport({ open, handleClose }) {
           choix.
         </Typography>
         <Grid spacing={4} container className={classes.grid}>
-          <Grid item sm={6} xs={12}>
-            <DatePicker
-              required
-              label="Mois de début"
-              value={minDate}
-              format="MMMM yyyy"
-              fullWidth
-              onChange={e => {
-                setMinDate(e);
-              }}
-              clearable
-              cancelLabel={null}
-              clearLabel="Annuler"
-              autoOk
-              disableFuture
-              inputVariant="outlined"
-              animateYearScrolling
-              error={!!dateRangeError}
-              helperText={dateRangeError}
-              openTo={"month"}
-              views={["year", "month"]}
-            />
-          </Grid>
-          <Grid item sm={6} xs={12}>
-            <DatePicker
-              required
-              label="Mois de fin"
-              value={maxDate}
-              format="MMMM yyyy"
-              fullWidth
-              onChange={e => {
-                console.log(e);
-                setMaxDate(e);
-              }}
-              clearable
-              cancelLabel={null}
-              clearLabel="Annuler"
-              autoOk
-              disableFuture
-              inputVariant="outlined"
-              animateYearScrolling
-              error={!!dateRangeError}
-              helperText={dateRangeError}
-              openTo={"month"}
-              views={["year", "month"]}
-            />
-          </Grid>
+          <DateOrDateTimeRangeSelectionContext
+            start={minDate}
+            setStart={setMinDate}
+            end={maxDate}
+            setEnd={setMaxDate}
+          >
+            <Grid item sm={6} xs={12}>
+              <DatePicker
+                required
+                label="Mois de début"
+                value={minDate}
+                format="MMMM yyyy"
+                fullWidth
+                onChange={e => {
+                  setMinDate(e);
+                }}
+                cancelLabel={null}
+                autoOk
+                disableFuture
+                inputVariant="outlined"
+                animateYearScrolling
+                error={!!dateRangeError}
+                helperText={dateRangeError}
+                openTo={"month"}
+                views={["year", "month"]}
+              />
+            </Grid>
+            <Grid item sm={6} xs={12}>
+              <DatePicker
+                required
+                label="Mois de fin"
+                value={maxDate}
+                format="MMMM yyyy"
+                fullWidth
+                onChange={e => {
+                  setMaxDate(e);
+                }}
+                cancelLabel={null}
+                autoOk
+                disableFuture
+                inputVariant="outlined"
+                animateYearScrolling
+                error={!!dateRangeError}
+                helperText={dateRangeError}
+                openTo={"month"}
+                views={["year", "month"]}
+              />
+            </Grid>
+          </DateOrDateTimeRangeSelectionContext>
         </Grid>
       </DialogContent>
       <CustomDialogActions>
